requirements.txt
python-dotenv==1.0.1                    # Reads .env files
langchain==0.2.2                        # Core LangChain
langchain-community==0.2.3             # Integrations
langchain-openai==0.1.8                # OpenAI integration for LangChain

unstructured==0.10.19                  # Document loader (latest Python 3.10-compatible)

chromadb==0.5.0                        # Vector store (Chroma)
openai==1.31.1                         # OpenAI client
tiktoken==0.7.0                        # Tokenizer

# Note: For markdown parsing, install this separately after all base deps:
# pip install "unstructured[md]"
